chore:removed crystal-kyber & update commented changes
