+++ app/navigators/SettingStack.tsx
@@ -6,8 +6,11 @@ import { DeviceEventEmitter } from 'react-native'
 import { EventTypes } from '../constants'
 import { useConfiguration } from '../contexts/configuration'
 import { useTheme } from '../contexts/theme'
+import CreateWallet from '../screens/CreateWallet'
 import ExportWallet from '../screens/ExportWallet'
 import ExportWalletConfirmation from '../screens/ExportWalletConfirmation'
+import ImportSuccess from '../screens/ImportSuccess'
+import ImportWalletVerify from '../screens/ImportWalletConfirmation'
 import Language from '../screens/Language'
 import Onboarding from '../screens/Onboarding'
 import { createCarouselStyle } from '../screens/OnboardingPages'
@@ -112,6 +115,17 @@ const SettingStack: React.FC = () => {
           />
         )}
       </Stack.Screen>
+      <Stack.Screen
+        name={Screens.ImportWalletVerify}
+        component={ImportWalletVerify}
+        options={{ title: t('Screens.VerifyPhrase'), headerBackTestID: testIdWithKey('Back') }}
+      />
+      <Stack.Screen name={Screens.ImportSuccess} component={ImportSuccess} options={{ title: t('Screens.Success') }} />
+      <Stack.Screen
+        name={Screens.WalletOptions}
+        component={CreateWallet}
+        options={{ title: t('Screens.CreateWallet') }}
+      />
     </Stack.Navigator>
   )
 }
