+++ package.json
@@ -47,7 +47,6 @@
     "i18next": "^23.4.1",
     "lodash.flatten": "^4.4.0",
     "lodash.startcase": "^4.4.0",
-    "md5": "^2.3.0",
     "moment": "^2.29.4",
     "patch-package": "^8.0.0",
     "query-string": "^8.1.0",
@@ -86,7 +85,6 @@
     "react-native-webview": "^13.3.1",
     "react-native-zip-archive": "6.0.9",
     "rn-fetch-blob": "^0.12.0",
-    "sha3": "^2.1.4",
     "uuid": "^9.0.0",
     "stream-browserify": "^1.0.0",
     "readable-stream": "1.0.33",
